+++ src/lib.rs
@@ -1,10 +1,4 @@
-//! The `grok` library allows you to quickly parse and match potentially unstructured data
-//! into a structed result. It is especially helpful when parsing logfiles of all kinds. This
-//! [Rust](http://rust-lang.org) version is mainly a port from the
-//! [java version](https://github.com/thekrakken/java-grok)
-//! which in drew inspiration from the original
-//! [ruby version](https://github.com/logstash-plugins/logstash-filter-grok).
-#![doc(html_root_url = "https://docs.rs/grok/2.0.0")]
+#![doc = include_str!("../README.md")]
 
 include!(concat!(env!("OUT_DIR"), "/default_patterns.rs"));
 
@@ -31,24 +25,34 @@ mod pattern_parser;
 // 3. onig
 // 3. regex
 
+#[derive(Clone, Copy, PartialEq, Eq, Debug)]
+#[allow(unused)]
+pub(crate) enum Engine {
+    Pcre2,
+    FancyRegex,
+    Onig,
+    Regex,
+}
+
 #[doc(hidden)]
 #[cfg(feature = "pcre2")]
 use pcre2::{
     Pcre2Matches as MatchesInner, Pcre2MatchesIter as MatchesIterInner,
-    Pcre2Pattern as InnerPattern,
+    Pcre2Pattern as InnerPattern, ENGINE,
 };
 
 #[doc(hidden)]
 #[cfg(all(not(feature = "pcre2"), feature = "fancy-regex"))]
 use fancy_regex::{
     FancyRegexMatches as MatchesInner, FancyRegexMatchesIter as MatchesIterInner,
-    FancyRegexPattern as InnerPattern,
+    FancyRegexPattern as InnerPattern, ENGINE,
 };
 
 #[doc(hidden)]
 #[cfg(all(not(feature = "pcre2"), not(feature = "fancy-regex"), feature = "onig"))]
 use onig::{
     OnigMatches as MatchesInner, OnigMatchesIter as MatchesIterInner, OnigPattern as InnerPattern,
+    ENGINE,
 };
 
 #[doc(hidden)]
@@ -60,7 +64,7 @@ use onig::{
 ))]
 use regex::{
     RegexMatches as MatchesInner, RegexMatchesIter as MatchesIterInner,
-    RegexPattern as InnerPattern,
+    RegexPattern as InnerPattern, ENGINE,
 };
 
 use crate::pattern_parser::{grok_split, GrokComponent};
@@ -68,6 +72,7 @@ use crate::pattern_parser::{grok_split, GrokComponent};
 /// The `Pattern` represents a compiled regex, ready to be matched against arbitrary text.
 pub struct Pattern {
     inner: InnerPattern,
+    extracts: HashMap<String, String>,
     #[cfg(test)]
     text: String,
 }
@@ -76,10 +81,15 @@ impl Pattern {
     /// Creates a new pattern from a raw regex string and an alias map to identify the
     /// fields properly.
     #[inline(always)]
-    fn new(regex: &str, alias: HashMap<String, String>) -> Result<Self, Error> {
+    fn new(
+        regex: &str,
+        alias: HashMap<String, String>,
+        extracts: HashMap<String, String>,
+    ) -> Result<Self, Error> {
         let inner = InnerPattern::new(regex, &alias)?;
         Ok(Self {
             inner,
+            extracts,
             #[cfg(test)]
             text: regex.to_string(),
         })
@@ -90,6 +100,7 @@ impl Pattern {
     pub fn match_against<'a>(&'a self, text: &'a str) -> Option<Matches<'a>> {
         Some(Matches {
             inner: self.inner.match_against(text)?,
+            pattern: self,
         })
     }
 
@@ -98,11 +109,42 @@ impl Pattern {
     pub fn capture_names(&self) -> impl Iterator<Item = &str> {
         self.inner.capture_names()
     }
+
+    /// Returns the extract for the given name.
+    #[inline(always)]
+    pub fn get_extract(&self, name: &str) -> Option<&str> {
+        self.extracts.get(name).map(|s| s.as_str())
+    }
+}
+
+impl std::fmt::Debug for Pattern {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        if f.alternate() {
+            f.debug_struct("Pattern")
+                .field("regex", &self.inner.regex)
+                .field("extracts", &self.extracts)
+                .field("capture_names", &self.capture_names().collect::<Vec<_>>())
+                .finish()
+        } else {
+            f.debug_struct("Pattern")
+                .field("regex", &self.inner.regex)
+                .field(
+                    "extracts",
+                    &format!("{{ {:?} extract(s) }}", self.extracts.len()),
+                )
+                .field(
+                    "capture_names",
+                    &format!("{{ {:?} capture(s) }}", self.capture_names().count()),
+                )
+                .finish()
+        }
+    }
 }
 
 /// The `Matches` represent matched results from a `Pattern` against a provided text.
 pub struct Matches<'a> {
     inner: MatchesInner<'a>,
+    pattern: &'a Pattern,
 }
 
 impl<'a> Matches<'a> {
@@ -127,8 +169,15 @@ impl<'a> Matches<'a> {
     /// Returns the number of matches.
     #[cfg(test)]
     #[inline(always)]
+    #[allow(clippy::len_without_is_empty)]
     pub fn len(&self) -> usize {
         self.iter().count()
+    }
+
+    /// Returns the pattern that was used to match this `Matches` instance.
+    #[inline(always)]
+    pub fn pattern(&self) -> &Pattern {
+        self.pattern
     }
 }
 
@@ -178,75 +227,127 @@ pub fn patterns<'a>() -> &'a [(&'a str, &'a str)] {
     PATTERNS
 }
 
+/// Grok pattern parser.
+///
+/// This API is currently unstable and may be subject to change.
+pub mod parser {
+    pub use crate::pattern_parser::*;
+}
+
 /// The `Grok` struct is the main entry point into using this library.
 #[derive(Debug)]
 pub struct Grok {
-    patterns: BTreeMap<String, String>,
+    #[allow(unused)]
+    engine: Engine,
+    patterns: BTreeMap<Cow<'static, str>, Cow<'static, str>>,
 }
 
 impl Grok {
     /// Creates a new `Grok` instance with no patterns.
-    pub fn empty() -> Self {
-        Grok {
+    pub const fn empty() -> Self {
+        Self {
+            engine: ENGINE,
             patterns: BTreeMap::new(),
         }
     }
 
     /// Creates a new `Grok` instance and loads all the default patterns.
+    ///
+    /// For more information, see the [`mod@patterns`] module.
     pub fn with_default_patterns() -> Self {
-        let mut grok = Grok::empty();
-        for &(key, value) in PATTERNS {
-            grok.add_pattern(String::from(key), String::from(value));
+        Self {
+            engine: ENGINE,
+            patterns: BTreeMap::from_iter(PATTERNS_COW.iter().cloned()),
         }
-        grok
     }
 
-    /// Adds a custom pattern.
+    /// Adds a custom grok pattern.
+    ///
+    /// A grok pattern is a standard regular expression string with grok pattern
+    /// placeholders embedded in it.
+    ///
+    /// The grok pattern placeholders are of the form
+    /// `%{name:alias:extract=definition}`, where `name` is the name of the
+    /// pattern, `alias` is the alias of the pattern, `extract` is the extract
+    /// of the pattern, and `definition` is the definition of the pattern.
+    ///
+    /// - `name` is the name of the pattern and is required. It may contain any
+    ///   alphanumeric character, or `_`.
+    /// - `alias` is the alias of the pattern and is optional. It may contain
+    ///   any alphanumeric character, or any of `_-[].`. If extract is provided,
+    ///   `alias` may be empty.
+    /// - `extract` is the extract of the pattern and is optional. It may
+    ///   contain any alphanumeric character, or any of `_-[].`.
+    /// - `definition` is the definition of the pattern and is optional. It may
+    ///   contain any character other than `{` or `}`.
+    ///
+    /// A literal `%` character may appear in a grok pattern as long as it is
+    /// not followed by `{`. You can surround the percent with grouped
+    /// parentheses `(%){..}`, a non-capturing group `(?:%){..}`, or use the
+    /// `\x25` escape sequence, ie: `\x25{..}`.
     pub fn add_pattern<S: Into<String>>(&mut self, name: S, pattern: S) {
-        self.patterns.insert(name.into(), pattern.into());
+        self.patterns
+            .insert(Cow::Owned(name.into()), Cow::Owned(pattern.into()));
     }
 
     /// Compiles the given pattern, making it ready for matching.
     ///
     /// Specify `with_alias_only` to only include the aliases in the matches
-    /// rather that all named patterns.
-    pub fn compile(&mut self, pattern: &str, with_alias_only: bool) -> Result<Pattern, Error> {
+    /// rather that all named patterns. This may result in a more efficient
+    /// compiled pattern.
+    pub fn compile(&self, pattern: &str, with_alias_only: bool) -> Result<Pattern, Error> {
+        let (named_regex, aliases, extracts) = self.compile_regex(pattern, with_alias_only)?;
+        if named_regex.is_empty() {
+            Err(Error::CompiledPatternIsEmpty(pattern.into()))
+        } else {
+            Pattern::new(&named_regex, aliases, extracts)
+        }
+    }
+
+    fn compile_regex(
+        &self,
+        pattern: &str,
+        with_alias_only: bool,
+    ) -> Result<(String, HashMap<String, String>, HashMap<String, String>), Error> {
         let mut named_regex = String::with_capacity(pattern.len() * 4);
         let mut aliases: HashMap<String, String> = HashMap::new();
         let mut aliases_extra: HashMap<String, usize> = HashMap::new();
+        let mut extracts: HashMap<String, String> = HashMap::new();
 
         let mut pattern_stack = Vec::with_capacity(16);
 
-        pattern_stack.push(grok_split(pattern));
+        pattern_stack.push((grok_split(pattern), BTreeMap::new()));
         let mut index = 0;
-        let mut pattern_overrides = BTreeMap::new();
 
-        while let Some(mut it) = pattern_stack.pop() {
+        while let Some((mut it, pattern_overrides)) = pattern_stack.pop() {
             if let Some(next) = it.next() {
-                pattern_stack.push(it);
+                pattern_stack.push((it, pattern_overrides));
                 use GrokComponent::*;
                 match next {
                     GrokPattern {
                         name,
                         alias,
-                        #[expect(unused)]
-                        capture, // TODO: use
+                        extract,
                         definition,
                         ..
                     } => {
                         if !definition.is_empty() {
                             // We can cleverly reborrow the definition here because we know that
                             // the lifetime is compatible.
-                            pattern_stack.push(grok_split(definition));
-                            pattern_overrides.insert(name.to_string(), definition);
-                        } else if let Some(pattern) = pattern_overrides.get(name) {
+                            pattern_stack
+                                .last_mut()
+                                .unwrap()
+                                .1
+                                .insert(name.to_string(), definition);
+                            pattern_stack.push((grok_split(definition), BTreeMap::new()));
+                        } else if let Some(pattern) = pattern_stack.last().unwrap().1.get(name) {
                             // Again, cleverly reborrow the pattern
-                            pattern_stack.push(grok_split(*pattern));
+                            pattern_stack.push((grok_split(*pattern), BTreeMap::new()));
                         } else {
                             let Some(pattern) = self.patterns.get(name) else {
                                 return Err(Error::DefinitionNotFound(name.to_string()));
                             };
-                            pattern_stack.push(grok_split(pattern));
+                            pattern_stack.push((grok_split(pattern), BTreeMap::new()));
                         }
 
                         if with_alias_only && alias.is_empty() {
@@ -272,6 +373,9 @@ impl Grok {
                                 )));
                             }
 
+                            if !extract.is_empty() {
+                                extracts.insert(key.clone(), extract.to_string());
+                            }
                             aliases.insert(match_name.clone(), key);
 
                             named_regex.push_str("(?<");
@@ -296,12 +400,7 @@ impl Grok {
         }
 
         named_regex.pop();
-
-        if named_regex.is_empty() {
-            Err(Error::CompiledPatternIsEmpty(pattern.into()))
-        } else {
-            Pattern::new(&named_regex, aliases)
-        }
+        Ok((named_regex, aliases, extracts))
     }
 }
 
@@ -433,7 +532,7 @@ mod tests {
     #[test]
     fn test_from_iter() {
         let patterns = [("USERNAME", r"[a-zA-Z0-9._-]+")];
-        let mut grok = Grok::from_iter(patterns.into_iter());
+        let grok = Grok::from_iter(patterns);
         let pattern = grok
             .compile("%{USERNAME}", false)
             .expect("Error while compiling!");
@@ -450,7 +549,7 @@ mod tests {
 
     #[test]
     fn test_from() {
-        let mut grok = Grok::from([("USERNAME", r"[a-zA-Z0-9._-]+")]);
+        let grok = Grok::from([("USERNAME", r"[a-zA-Z0-9._-]+")]);
         let pattern = grok
             .compile("%{USERNAME}", false)
             .expect("Error while compiling!");
@@ -537,7 +636,7 @@ mod tests {
             .match_against("hello! 5E:FF:56:A2:AF:15 what?")
             .expect("No matches found!");
         assert_eq!("5E:FF:56:A2:AF:15", matches.get("MAC").unwrap());
-        assert_eq!(true, pattern.match_against("5E:FF").is_none());
+        assert!(pattern.match_against("5E:FF").is_none());
     }
 
     #[test]
@@ -584,7 +683,7 @@ mod tests {
             .match_against("hello! 5E:FF:56:A2:AF:15 what?")
             .expect("No matches found!");
         assert_eq!("5E:FF:56:A2:AF:15", matches.get("macaddr").unwrap());
-        assert_eq!(true, pattern.match_against("5E:FF").is_none());
+        assert!(pattern.match_against("5E:FF").is_none());
     }
 
     #[test]
@@ -653,10 +752,12 @@ mod tests {
         assert_eq!(4, found);
     }
 
-    #[cfg(not(feature = "regex"))]
     #[test]
     fn test_loaded_default_patterns() {
-        let mut grok = Grok::with_default_patterns();
+        if ENGINE == Engine::Regex {
+            return;
+        }
+        let grok = Grok::with_default_patterns();
         let pattern = grok
             .compile("%{DAY} %{MONTH} %{YEAR}", false)
             .expect("Error while compiling!");
@@ -670,13 +771,15 @@ mod tests {
         assert_eq!(None, matches.get("unknown"));
     }
 
-    #[cfg(not(feature = "regex"))]
     #[test]
     fn test_compilation_of_all_default_patterns() {
-        let mut grok = Grok::default();
+        if ENGINE == Engine::Regex {
+            return;
+        }
+        let grok = Grok::default();
         let mut num_checked = 0;
         let mut errors = vec![];
-        for &(key, _) in PATTERNS {
+        for &(key, _) in patterns() {
             let pattern = format!("%{{{}}}", key);
             match grok.compile(&pattern, false) {
                 Ok(_) => (),
@@ -695,7 +798,7 @@ mod tests {
 
     #[test]
     fn test_adhoc_pattern() {
-        let mut grok = Grok::default();
+        let grok = Grok::default();
         let pattern = grok
             .compile(r"\[(?<threadname>[^\]]+)\]", false)
             .expect("Error while compiling!");
@@ -709,7 +812,7 @@ mod tests {
 
     #[test]
     fn test_adhoc_pattern_in_iter() {
-        let mut grok = Grok::default();
+        let grok = Grok::default();
         let pattern = grok
             .compile(r"\[(?<threadname>[^\]]+)\]", false)
             .expect("Error while compiling!");
@@ -730,7 +833,7 @@ mod tests {
     /// If multiple captures have the same name, the last one wins.
     #[test]
     fn test_adhoc_pattern_conflict() {
-        let mut grok = Grok::with_default_patterns();
+        let grok = Grok::with_default_patterns();
         let pattern = grok
             .compile(r"(?<capture>\w+) %{GREEDYDATA:capture}", true)
             .unwrap();
@@ -741,7 +844,7 @@ mod tests {
 
     #[test]
     fn test_capture_repeat() {
-        let mut grok = Grok::with_default_patterns();
+        let grok = Grok::with_default_patterns();
         let pattern = grok.compile(r"%{INT}{1,3}", false).unwrap();
         let matches = pattern.match_against("+1+2+3").unwrap();
         assert_eq!("+3", matches.get("INT").unwrap());
@@ -749,7 +852,7 @@ mod tests {
 
     #[test]
     fn test_pattern_with_definition() {
-        let mut grok = Grok::default();
+        let grok = Grok::default();
         let pattern = grok
             .compile(r"%{NEW_PATTERN:first=\w+} %{NEW_PATTERN:second}", false)
             .unwrap();
@@ -776,10 +879,37 @@ mod tests {
         assert_eq!(expected, actual);
     }
 
-    #[cfg(not(feature = "regex"))]
+    #[test]
+    fn test_capture_names_with_type() {
+        let mut grok = Grok::empty();
+        grok.add_pattern("USERNAME", r"[a-zA-Z0-9._-]+");
+        grok.add_pattern("USER", r"%{USERNAME::text}");
+        let pattern = grok
+            .compile("%{USER:usr:text}", true)
+            .expect("Error while compiling!");
+        eprintln!("{pattern:#?}");
+
+        let actual = pattern.capture_names().collect::<Vec<_>>();
+        assert_eq!(vec!["usr"], actual);
+
+        assert_eq!(Some("text"), pattern.get_extract("usr"));
+        assert_eq!(None, pattern.get_extract("USERNAME"));
+        assert_eq!(None, pattern.get_extract("USER"));
+        assert_eq!(None, pattern.get_extract("doesn't exist"));
+
+        let pattern = grok
+            .compile("%{USER:usr:text}", false)
+            .expect("Error while compiling!");
+        eprintln!("{pattern:#?}");
+    }
+
     #[test]
     fn test_capture_error() {
-        let mut grok = Grok::with_default_patterns();
+        if ENGINE == Engine::Regex {
+            return;
+        }
+
+        let grok = Grok::with_default_patterns();
         let pattern = grok
             .compile("Path: %{PATH}$", false)
             .expect("Error while compiling!");
@@ -789,10 +919,13 @@ mod tests {
         assert!(matches.is_none());
     }
 
-    #[cfg(not(feature = "regex"))]
     #[test]
     fn test_match_deep_patterns() {
-        let mut grok = Grok::with_default_patterns();
+        if ENGINE == Engine::Regex {
+            return;
+        }
+
+        let grok = Grok::with_default_patterns();
         let pattern = grok
             .compile("%{BACULA_LOGLINE}", false)
             .expect("Error while compiling!");
@@ -837,10 +970,13 @@ mod tests {
         eprintln!("{:?}", matches);
     }
 
-    #[cfg(not(feature = "regex"))]
     #[test]
     fn test_compile_deep_patterns() {
-        let mut grok = Grok::with_default_patterns();
+        if ENGINE == Engine::Regex {
+            return;
+        }
+
+        let grok = Grok::with_default_patterns();
         let pattern = grok
             .compile("%{BACULA_LOGLINE}", false)
             .expect("Error while compiling!");